ingest-rss/app/requirements.txt
langchain==0.1.13
langchain-openai==0.1.1
pinecone-client==3.2.1
langchain-community==0.0.29
langchain-pinecone==0.0.3
feedparser==6.0.11
python-dotenv==1.0.1
newspaper3k==0.2.8